# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Restful Provisioning Protocol (RPP) ",
    "description": "A Modern Domain Registry API",
    "version": "0.0.1"
  },
  "paths": {
...
    "/organisations/": {
      "post": {
        "tags": [
          "Organisations"
        ],
        "summary": "Create Organisation",
        "operationId": "do_create_organisations__post",
        "security": [
          {
            "HTTPBasic": []
          },
          {
            "HTTPBasic": []
          }
        ],
        "parameters": [
          {
            "name": "rpp-cl-trid",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Rpp-Cl-Trid"
            }
          },
          {
            "name": "session_id",
            "in": "cookie",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrganisationCreateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponseModel"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/organisations/{organisation}": {
      "get": {
        "tags": [
          "Organisations"
        ],
        "summary": "Get Organisation Info",
        "operationId": "do_info_organisations__organisation__get",
        "security": [
          {
            "HTTPBasic": []
          },
          {
            "HTTPBasic": []
          }
        ],
        "parameters": [
          {
            "name": "organisation",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Organisation"
            }
          },
          {
            "name": "rpp-authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Rpp-Authorization"
            }
          },
          {
            "name": "rpp-cl-trid",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Rpp-Cl-Trid"
            }
          },
          {
            "name": "session_id",
            "in": "cookie",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponseModel"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Organisations"
        ],
        "summary": "Delete Organisation",
        "operationId": "do_delete_organisations__organisation__delete",
        "security": [
          {
            "HTTPBasic": []
          },
          {
            "HTTPBasic": []
          }
        ],
        "parameters": [
          {
            "name": "organisation",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Organisation"
            }
          },
          {
            "name": "rpp-cl-trid",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Rpp-Cl-Trid"
            }
          },
          {
            "name": "session_id",
            "in": "cookie",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Organisations"
        ],
        "summary": "Update Organisation",
        "operationId": "do_update_organisations__organisation__patch",
        "security": [
          {
            "HTTPBasic": []
          },
          {
            "HTTPBasic": []
          }
        ],
        "parameters": [
          {
            "name": "organisation",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Organisation"
            }
          },
          {
            "name": "rpp-cl-trid",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Rpp-Cl-Trid"
            }
          },
          {
            "name": "session_id",
            "in": "cookie",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrganisationUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponseModel"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Address": {
        "properties": {
          "components": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/AddressComponent"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Components"
          },
          "countryCode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Countrycode"
          },
          "coordinates": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Coordinates"
          },
          "full": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Full"
          },
          "contexts": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "boolean"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contexts"
          }
        },
        "type": "object",
        "title": "Address"
      },
      "AddressComponent": {
        "properties": {
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "value": {
            "type": "string",
            "title": "Value"
          }
        },
        "type": "object",
        "required": [
          "kind",
          "value"
        ],
        "title": "AddressComponent"
      },
      "Addresses": {
        "additionalProperties": {
          "$ref": "#/components/schemas/Address"
        },
        "type": "object",
        "title": "Addresses"
      },
...
      "BaseResponseModel": {
        "properties": {
          "@type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "@Type"
          },
          "result": {
            "items": {
              "$ref": "#/components/schemas/ResultModel"
            },
            "type": "array",
            "title": "Result"
          },
          "messages": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MessageQueueModel"
              },
              {
                "type": "null"
              }
            ]
          },
          "resData": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Resdata"
          },
          "extension": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Extension"
          }
        },
        "type": "object",
        "required": [
          "result"
        ],
        "title": "BaseResponseModel"
      },
      "Card": {
        "properties": {
          "@type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "@Type",
            "default": "Card"
          },
          "version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Version",
            "default": "2.0"
          },
          "rpp.ietf.org:id": {
            "type": "string",
            "title": "Rpp.Ietf.Org:Id"
          },
          "name": {
            "$ref": "#/components/schemas/Name"
          },
          "organizations": {
            "additionalProperties": {
              "$ref": "#/components/schemas/Organization"
            },
            "type": "object",
            "title": "Organizations"
          },
          "addresses": {
            "$ref": "#/components/schemas/Addresses"
          },
          "phones": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Phones"
              },
              {
                "type": "null"
              }
            ]
          },
          "emails": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Emails"
              },
              {
                "type": "null"
              }
            ]
          },
          "int_": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Int ",
            "default": false
          },
          "rpp.ietf.org:legalForm": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SidnLegalForm"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "rpp.ietf.org:id",
          "name",
          "addresses"
        ],
        "title": "Card"
      },
...
      "Email": {
        "properties": {
          "address": {
            "type": "string",
            "title": "Address"
          }
        },
        "type": "object",
        "required": [
          "address"
        ],
        "title": "Email"
      },
      "Emails": {
        "additionalProperties": {
          "$ref": "#/components/schemas/Email"
        },
        "type": "object",
        "title": "Emails"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LinkType": {
        "properties": {
          "uri": {
            "type": "string",
            "title": "Uri"
          }
        },
        "type": "object",
        "required": [
          "uri"
        ],
        "title": "LinkType"
      },
      "LinksType": {
        "additionalProperties": {
          "$ref": "#/components/schemas/LinkType"
        },
        "type": "object",
        "title": "LinksType"
      },
      "MessageQueueModel": {
        "properties": {
          "count": {
            "type": "integer",
            "title": "Count"
          },
          "id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "qDate": {
            "type": "string",
            "format": "date-time",
            "title": "Qdate"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "count"
        ],
        "title": "MessageQueueModel"
      },
      "Name": {
        "properties": {
          "full": {
            "type": "string",
            "title": "Full"
          },
          "components": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/NameComponent"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Components"
          }
        },
        "type": "object",
        "required": [
          "full"
        ],
        "title": "Name"
      },
      "NameComponent": {
        "properties": {
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "value": {
            "type": "string",
            "title": "Value"
          }
        },
        "type": "object",
        "required": [
          "kind",
          "value"
        ],
        "title": "NameComponent"
      },
...
      "OrganisationCard": {
        "properties": {
          "@type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "@Type",
            "default": "Card"
          },
          "version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Version",
            "default": "2.0"
          },
          "rpp.ietf.org:id": {
            "type": "string",
            "title": "Rpp.Ietf.Org:Id"
          },
          "name": {
            "$ref": "#/components/schemas/Name"
          },
          "organizations": {
            "additionalProperties": {
              "$ref": "#/components/schemas/Organization"
            },
            "type": "object",
            "title": "Organizations"
          },
          "addresses": {
            "$ref": "#/components/schemas/Addresses"
          },
          "phones": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Phones"
              },
              {
                "type": "null"
              }
            ]
          },
          "emails": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Emails"
              },
              {
                "type": "null"
              }
            ]
          },
          "int_": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Int ",
            "default": false
          },
          "rpp.ietf.org:legalForm": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SidnLegalForm"
              },
              {
                "type": "null"
              }
            ]
          },
          "links": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LinksType"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "rpp.ietf.org:id",
          "name",
          "addresses"
        ],
        "title": "OrganisationCard"
      },
      "OrganisationContact": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type"
          },
          "id": {
            "type": "string",
            "title": "Id"
          }
        },
        "type": "object",
        "required": [
          "type",
          "id"
        ],
        "title": "OrganisationContact"
      },
      "OrganisationCreateRequest": {
        "properties": {
          "@type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "@Type"
          },
          "id": {
            "type": "string",
            "title": "Id"
          },
          "role": {
            "$ref": "#/components/schemas/OrganisationRole"
          },
          "status": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/OrganisationStatusType"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          },
          "parentId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parentid"
          },
          "contacts": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/OrganisationContact"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contacts"
          },
          "cards": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/OrganisationCard"
                },
                "type": "array",
                "maxItems": 2
              },
              {
                "type": "null"
              }
            ],
            "title": "Cards"
          }
        },
        "type": "object",
        "required": [
          "id",
          "role"
        ],
        "title": "OrganisationCreateRequest"
      },
      "OrganisationRole": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type"
          },
          "status": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/RoleStatusType"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          },
          "roleID": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Roleid"
          }
        },
        "type": "object",
        "required": [
          "type"
        ],
        "title": "OrganisationRole"
      },
      "OrganisationStatusType": {
        "type": "string",
        "enum": [
          "ok",
          "hold",
          "terminated",
          "clientDeleteProhibited",
          "clientUpdateProhibited",
          "clientLinkProhibited",
          "linked",
          "pendingCreate",
          "pendingUpdate",
          "pendingDelete",
          "serverDeleteProhibited",
          "serverUpdateProhibited",
          "serverLinkProhibited"
        ],
        "title": "OrganisationStatusType"
      },
      "OrganisationUpdateAddOrRemove": {
        "properties": {
          "status": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "OrganisationUpdateAddOrRemove"
      },
      "OrganisationUpdateChange": {
        "properties": {
          "contact": {
            "items": {
              "$ref": "#/components/schemas/Card"
            },
            "type": "array",
            "title": "Contact"
          }
        },
        "type": "object",
        "required": [
          "contact"
        ],
        "title": "OrganisationUpdateChange"
      },
      "OrganisationUpdateRequest": {
        "properties": {
          "@type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "@Type"
          },
          "add": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/OrganisationUpdateAddOrRemove"
              },
              {
                "type": "null"
              }
            ]
          },
          "remove": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/OrganisationUpdateAddOrRemove"
              },
              {
                "type": "null"
              }
            ]
          },
          "change": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/OrganisationUpdateChange"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "title": "OrganisationUpdateRequest"
      },
      "Organization": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "Organization"
      },
...
      "Phone": {
        "properties": {
          "features": {
            "$ref": "#/components/schemas/PhoneFeatures"
          },
          "number": {
            "type": "string",
            "title": "Number"
          }
        },
        "type": "object",
        "required": [
          "features",
          "number"
        ],
        "title": "Phone"
      },
      "PhoneFeatures": {
        "properties": {
          "voice": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Voice"
          },
          "fax": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fax"
          }
        },
        "type": "object",
        "title": "PhoneFeatures"
      },
      "Phones": {
        "additionalProperties": {
          "$ref": "#/components/schemas/Phone"
        },
        "type": "object",
        "title": "Phones"
      },
      "ResultModel": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "lang": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lang"
          },
          "value": {
            "anyOf": [
              {
                "items": {},
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Value"
          },
          "extValue": {
            "anyOf": [
              {
                "items": {},
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Extvalue"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ResultModel"
      },
      "RoleStatusType": {
        "type": "string",
        "enum": [
          "ok",
          "clientLinkProhibited",
          "linked",
          "serverLinkProhibited"
        ],
        "title": "RoleStatusType"
      },
...
      "SidnLegalForm": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "number": {
            "type": "string",
            "title": "Number"
          }
        },
        "type": "object",
        "required": [
          "name",
          "number"
        ],
        "title": "SidnLegalForm"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBasic": {
        "type": "http",
        "scheme": "basic"
      }
    }
  }
}